chore: remove common file
